# fractalx-output/fractalx-gateway/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "FractalX Generated API"
  version: "1.0.0"
  description: |
    Auto-generated by FractalX decomposition framework.
    All routes are proxied through the API Gateway on port 9999.
servers:
  - url: "http://localhost:9999"
    description: "API Gateway"
  - url: "http://localhost:8085"
    description: "Department Service (direct)"
  - url: "http://localhost:8081"
    description: "Employee Service (direct)"
  - url: "http://localhost:8083"
    description: "Leave Service (direct)"
  - url: "http://localhost:8082"
    description: "Payroll Service (direct)"
  - url: "http://localhost:8084"
    description: "Recruitment Service (direct)"
tags:
  - name: "department-service"
    description: "Department Service API"
  - name: "employee-service"
    description: "Employee Service API"
  - name: "leave-service"
    description: "Leave Service API"
  - name: "payroll-service"
    description: "Payroll Service API"
  - name: "recruitment-service"
    description: "Recruitment Service API"
  - name: "gateway"
    description: "Gateway management endpoints"
paths:
  /api/department:
    get:
      tags: ["department-service"]
      summary: "List Department Service resources"
      operationId: "department-service-list"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    post:
      tags: ["department-service"]
      summary: "Create Department Service resource"
      operationId: "department-service-create"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "201":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/department/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "Resource identifier"
        schema:
          type: string
    get:
      tags: ["department-service"]
      summary: "Get Department Service by ID"
      operationId: "department-service-get-by-id"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    put:
      tags: ["department-service"]
      summary: "Update Department Service by ID"
      operationId: "department-service-update"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    delete:
      tags: ["department-service"]
      summary: "Delete Department Service by ID"
      operationId: "department-service-delete"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: "Deleted successfully"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/employee:
    get:
      tags: ["employee-service"]
      summary: "List Employee Service resources"
      operationId: "employee-service-list"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    post:
      tags: ["employee-service"]
      summary: "Create Employee Service resource"
      operationId: "employee-service-create"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "201":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/employee/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "Resource identifier"
        schema:
          type: string
    get:
      tags: ["employee-service"]
      summary: "Get Employee Service by ID"
      operationId: "employee-service-get-by-id"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    put:
      tags: ["employee-service"]
      summary: "Update Employee Service by ID"
      operationId: "employee-service-update"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    delete:
      tags: ["employee-service"]
      summary: "Delete Employee Service by ID"
      operationId: "employee-service-delete"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: "Deleted successfully"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/leave:
    get:
      tags: ["leave-service"]
      summary: "List Leave Service resources"
      operationId: "leave-service-list"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    post:
      tags: ["leave-service"]
      summary: "Create Leave Service resource"
      operationId: "leave-service-create"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "201":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/leave/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "Resource identifier"
        schema:
          type: string
    get:
      tags: ["leave-service"]
      summary: "Get Leave Service by ID"
      operationId: "leave-service-get-by-id"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    put:
      tags: ["leave-service"]
      summary: "Update Leave Service by ID"
      operationId: "leave-service-update"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    delete:
      tags: ["leave-service"]
      summary: "Delete Leave Service by ID"
      operationId: "leave-service-delete"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: "Deleted successfully"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/payroll:
    get:
      tags: ["payroll-service"]
      summary: "List Payroll Service resources"
      operationId: "payroll-service-list"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    post:
      tags: ["payroll-service"]
      summary: "Create Payroll Service resource"
      operationId: "payroll-service-create"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "201":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/payroll/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "Resource identifier"
        schema:
          type: string
    get:
      tags: ["payroll-service"]
      summary: "Get Payroll Service by ID"
      operationId: "payroll-service-get-by-id"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    put:
      tags: ["payroll-service"]
      summary: "Update Payroll Service by ID"
      operationId: "payroll-service-update"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    delete:
      tags: ["payroll-service"]
      summary: "Delete Payroll Service by ID"
      operationId: "payroll-service-delete"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: "Deleted successfully"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/recruitment:
    get:
      tags: ["recruitment-service"]
      summary: "List Recruitment Service resources"
      operationId: "recruitment-service-list"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    post:
      tags: ["recruitment-service"]
      summary: "Create Recruitment Service resource"
      operationId: "recruitment-service-create"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "201":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /api/recruitment/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "Resource identifier"
        schema:
          type: string
    get:
      tags: ["recruitment-service"]
      summary: "Get Recruitment Service by ID"
      operationId: "recruitment-service-get-by-id"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
        - {}
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    put:
      tags: ["recruitment-service"]
      summary: "Update Recruitment Service by ID"
      operationId: "recruitment-service-update"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
    delete:
      tags: ["recruitment-service"]
      summary: "Delete Recruitment Service by ID"
      operationId: "recruitment-service-delete"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: "Deleted successfully"
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
  /actuator/health:
    get:
      tags: ["gateway"]
      summary: "Gateway health check"
      operationId: "gateway-health"
      responses:
        "200":
          description: "Gateway is healthy"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
components:
  schemas:
    GenericRequest:
      type: object
      additionalProperties: true
      description: "Request payload — replace with your concrete schema"
    GenericResponse:
      type: object
      additionalProperties: true
      description: "Response payload — replace with your concrete schema"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key